+++ CHANGES.md
@@ -1,8 +1,25 @@
 ## 0.3.0-SNAPSHOT
 
-* **BREAKING**: Removed type-level interceptors from ring-adapter.
+* **BREAKING**: Removed type-level interceptors from ring-adapter, use normal interceptors instead.
+* **BREAKING**: Ring request-parameters are now assoc-in'd (into `:data`) instead of deep-merging. For speed.
+* Handlers can be now be mounted to dispatcher root
 * Support for Context-based urls, thanks to [Wout Neirynck](https://github.com/wneirynck).
 * Data input schemas for apis can be vectors, fixes [#27](https://github.com/metosin/kekkonen/issues/27).
+* Use Pedestal-style interceptors, with `:name`, `:enter`, `:leave` and `:error`
+  * Extended to contain `:input` and `:output` schemas.
+* Interceptors are pre-compiled in all layers for simplicity and better perf.
+* Remove the following excess meta-data from handlers: 
+  * `:ns-meta`, `:all-meta`, `:handler-input` & `:user-input`
+* Remove `:interceptors` from the `Dispatcher`, as they are now precompiled into handlers
+* Interceptors can be `nil`, allowing conditional interceptors
+
+```clj
+(k/handler
+  {:name "fixture!"
+   :interceptors [[require-role :admin] (if-not env/dev-mode? log-it)]
+   :handle (fn [ctx] ...)})
+```
+
 * **BREAKING**: top-level swagger options are now in align to the compojure-api:
   * Fixes [#22](https://github.com/metosin/kekkonen/issues/22)
   * By default, `api`s don't bind swagger-spec & swagger-ui, use `:spec` & `:ui` options
@@ -37,10 +54,11 @@
 * updated dependencies:
 
 ```clj
-[prismatic/schema "1.1.1"] is available but we use "1.1.0"
+[prismatic/schema "1.1.2"] is available but we use "1.1.0"
 [prismatic/plumbing "0.5.3"] is available but we use "0.5.2"
-[metosin/ring-swagger "0.22.8"] is available but we use "0.22.6"
-[clj-http "3.1.0"] available but we use "2.1.0"
+[metosin/ring-http-response "0.7.0"] is available but we use "0.6.5"
+[metosin/ring-swagger "0.22.9"] is available but we use "0.22.6"
+[clj-http "2.2.0"] available but we use "2.1.0"
 ```
 
 ## 0.2.0 (29.3.2016)